\name{getFilterTypes}
\alias{getFilterTypes}
\title{Get Available Filter Types}
\description{
  Get available filter types.
}
\usage{
getFilterTypes()
}
\value{
  A character vector with available filter types.
}
\references{
  \bibshow{wordnet::Wallace:2007}
}
\author{Ingo Feinerer}
\examples{
getFilterTypes()
}
\keyword{file}
