\name{getFilterTypes}
\alias{getFilterTypes}
\title{Get available filter types}
\description{
  Get available filter types.
}
\usage{
getFilterTypes()
}
\value{
  A character vector with available filter types.
}
\references{
  M. Wallace (2007).
  Jawbone Java WordNet API.
  \url{http://mfwallace.googlepages.com/jawbone.html}
}
\author{Ingo Feinerer}
\examples{
getFilterTypes()
}
\keyword{file}
