\name{getFilterTypes}
\alias{getFilterTypes}
\title{Get Available Filter Types}
\description{
  Get available filter types.
}
\usage{
getFilterTypes()
}
\value{
  A character vector with available filter types.
}
\references{
  M. Wallace (2007).
  Jawbone Java WordNet API.
  \url{https://sites.google.com/site/mfwallace/jawbone}
}
\author{Ingo Feinerer}
\examples{
getFilterTypes()
}
\keyword{file}
